Allow sharing getSearchURL and fixURL methods across frontends
