+++ app/ui/shared/util/location-util.js
@@ -1,0 +1,28 @@
+/*
+Copyright 2016 Mozilla
+
+Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+this file except in compliance with the License. You may obtain a copy of the
+License at http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software distributed
+under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+CONDITIONS OF ANY KIND, either express or implied. See the License for the
+specific language governing permissions and limitations under the License.
+*/
+
+export function getSearchURL(query) {
+  // See https://dxr.mozilla.org/mozilla-central/source/browser/locales/en-US/searchplugins
+  return `https://www.duckduckgo.com/?q=${encodeURIComponent(query)}`;
+}
+
+export function fixURL(typed) {
+  if (typed.includes('://') || typed.trim().startsWith('data:')) {
+    return typed;
+  }
+
+  if (!typed.includes(' ') && typed.includes('.')) {
+    return `http://${typed}`;
+  }
+
+  return getSearchURL(typed);
+}
